Extract InitDataHelper so it's reusable in other specs
Extract InitDataHelper so it's reusable in other specs
